# Replacing the KPI section with React