--- a/babel.config.js
+++ b/babel.config.js
@@ -10,6 +10,7 @@ module.exports = function (api) {
           targets: { chrome: "58", ie: "11" },
         },
       ],
+      ["@babel/preset-react"],
     ],
   };
 };