--- a/package.json
+++ b/package.json
@@ -26,6 +26,8 @@
     "i18next": "^19.3.4",
     "i18next-browser-languagedetector": "^4.0.2",
     "loc-i18next": "^0.1.4",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
     "tippy.js": "^6.1.0",
     "twemoji": "^12.1.5",
     "whatwg-fetch": "^3.0.0"
@@ -33,6 +35,7 @@
   "devDependencies": {
     "@babel/cli": "^7.8.4",
     "@babel/preset-env": "^7.9.0",
+    "@babel/preset-react": "^7.10.4",
     "babel-loader": "^8.0.5",
     "babel-plugin-date-fns": "^2.0.0",
     "clean-webpack-plugin": "^3.0.0",
@@ -40,7 +43,10 @@
     "css-loader": "^3.4.1",
     "eslint": "^6.8.0",
     "eslint-config-prettier": "^6.10.1",
+    "eslint-config-react-app": "^5.2.1",
     "eslint-plugin-prettier": "^3.1.2",
+    "eslint-plugin-react": "^7.20.3",
+    "eslint-plugin-react-hooks": "^4.0.8",
     "file-loader": "^5.0.2",
     "html-webpack-plugin": "^3.2.0",
     "husky": "^4.2.3",