--- a/package.json
+++ b/package.json
@@ -26,6 +26,7 @@
     "i18next": "^19.3.4",
     "i18next-browser-languagedetector": "^4.0.2",
     "loc-i18next": "^0.1.4",
+    "prop-types": "^15.7.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "tippy.js": "^6.1.0",