--- a/src/components/KpiReact/linearGradient.js
+++ b/src/components/KpiReact/linearGradient.js
@@ -1,0 +1,16 @@
+const gradient = {
+  confirmed:
+    '<svg style="width:0;height:0;position:absolute;" aria-hidden="true" focusable="false"><linearGradient id="kpi-confirmed-chart-gradient" x2="0" y2="1"><stop offset="0%" stop-color="rgba(244,67,54, 0.6)" /><stop offset="90%" stop-color="rgba(244,67,54, 0)" /><stop offset="100%" stop-color="rgba(244,67,54, 0)" /></linearGradient></svg>',
+  recovered:
+    '<svg style="width:0;height:0;position:absolute;" aria-hidden="true" focusable="false"><linearGradient id="kpi-recovered-chart-gradient" x2="0" y2="1"><stop offset="0%" stop-color="rgba(24,118,211, 0.6)" /><stop offset="90%" stop-color="rgba(24,118,211, 0)" /><stop offset="100%" stop-color="rgba(24,118,211, 0)" /></linearGradient></svg>',
+  critical:
+    '<svg style="width:0;height:0;position:absolute;" aria-hidden="true" focusable="false"><linearGradient id="kpi-critical-chart-gradient" x2="0" y2="1"><stop offset="0%" stop-color="rgba(114,4,5, 0.6)" /><stop offset="90%" stop-color="rgba(114,4,5, 0)" /><stop offset="100%" stop-color="rgba(114,4,5, 0)" /></linearGradient></svg>',
+  deceased:
+    '<svg style="width:0;height:0;position:absolute;" aria-hidden="true" focusable="false"><linearGradient id="kpi-deceased-chart-gradient" x2="0" y2="1"><stop offset="0%" stop-color="rgba(54,72,79, 0.6)" /><stop offset="90%" stop-color="rgba(54,72,79, 0)" /><stop offset="100%" stop-color="rgba(54,72,79, 0)" /></linearGradient></svg>',
+  active:
+    '<svg style="width:0;height:0;position:absolute;" aria-hidden="true" focusable="false"><linearGradient id="kpi-active-chart-gradient" x2="0" y2="1"><stop offset="0%" stop-color="rgba(223,14,31, 0.6)" /><stop offset="90%" stop-color="rgba(223,14,31, 0)" /><stop offset="100%" stop-color="rgba(223,14,31, 0)" /></linearGradient></svg>',
+  tested:
+    '<svg style="width:0;height:0;position:absolute;" aria-hidden="true" focusable="false"><linearGradient id="kpi-tested-chart-gradient" x2="0" y2="1"><stop offset="0%" stop-color="rgba(164,173,192, 0.6)" /><stop offset="90%" stop-color="rgba(164,173,192, 0)" /><stop offset="100%" stop-color="rgba(164,173,192, 0)" /></linearGradient></svg>',
+};
+
+export default gradient;