--- a/src/index.html
+++ b/src/index.html
@@ -178,6 +178,8 @@
       </div>
     </section>
 
+    <div id="react-container"></div>
+
     <section id="national-charts" class="embed-hide">
       <section id="trend-chart-container" class="embed-hide">
         <h4 data-i18n="outbreak-spread-trend">Total Cases</h4>