// Please tag @reustle in your newly created PR
@reustle @liquidx 
// Please include a screenshot of any visual changes you've made
#237 
I wanted to give this a try and test out the performance of the page locally. I'd love to continue the discussion of migrating away from vanilla in favour of React or even NextJS for it's built-in server side rendering. 

Contents of this PR:
- added react, react-dom, and babel-react to start using React in this repo.
- added and configured eslint to include react and react-hooks.
- added react-i18next to allow for i18n hook to get current language and set translations within React.
- created a container component (KpiReact > index.js) which takes in the ddb data and constructs the kpi data to map over.
- created a presentational component (KpiReact > kpi.js) which fetches each individual graph and renders the kpi data.
- included a helper (KpiReact > linearGradient.js) to add the gradient overlay of the svg charts

I am aware that there is some layout jump at the beginning during the render phase. If anyone has a suggestion or recommendation to reduce that please let me know.

I've attached lighthouse reports for before and after the addition of React.
[lighthouse.zip](https://github.com/reustle/covid19japan/files/4978993/lighthouse.zip)

**Summary:** 
**Speed Index:** 1.0s to 0.5s
**Time to Interactive:** 8.4s to 6.5s
**Total Blocking Time:** 5,810ms to 4,970ms
**Cumulative Layout Shift:** 0.03s to 0.014s

**No visual change:**
<img width="1792" alt="Screen Shot 2020-07-27 at 3 59 43 PM" src="https://user-images.githubusercontent.com/11080410/88586355-4b5be300-d022-11ea-851b-46652e49fc59.png">

Thanks for reading!