@liquidx 2020-08-03
This is interesting, I'm not that familiar with React. One thing I noticed is that now, when the page first loads and the data is not yet loaded, the KPI boxes are not there.

Is there a way that has the original loading state to stay like it does in the current site?