@dshlass 2020-08-03
> This is interesting, I'm not that familiar with React. One thing I noticed is that now, when the page first loads and the data is not yet loaded, the KPI boxes are not there.
> 
> Is there a way that has the original loading state to stay like it does in the current site?

There is, but I am having some difficulty implementing it without a package - which I am hesitant to add keep the packages lean. 

I'd love some input on methods to render the kpi children to specific html elements rather than a parent node (render each KPI section individually to the target section in the HTML code). 

One solution I was looking into was to render the component at the start of the "DOMContentLoaded" event without data and use redux or another global state management tool to keep track of the changing of the global data values. The problem is that without a discussion on the migration strategy to React it could overcomplicate things fairly quickly. 

I've never converted an app from vanilla to React in pieces and I can see that is will be a fun challenge. I've got lots of time to play around with the implementation, I'd just love some input from another React dev on the matter.