@vlumi 2020-08-05
> > Is there a way that has the original loading state to stay like it does in the current site?
> 
> There is, but I am having some difficulty implementing it without a package - which I am hesitant to add keep the packages lean.

One quick hack would be to leave the original static DOM in place, so that it'll be shown until React replaces it with its own rendering. Reverting your changes to `src/index.html` seemed to work fairly well, though there is some wobbling and a quick flick when the kpi element is empty for a short moment...