@dshlass 2020-08-06
>  Reverting your changes to `src/index.html` seemed to work fairly well, though there is some wobbling and a quick flick when the kpi element is empty for a short moment...

Yeah I actually tried this as well but thought the site _felt_ better when it was just delayed loading rather than the flick/wobble. I am going to try to set a min-height for the kpi section and see if having the blank space without the page jump feels better.