@vlumi 2020-08-07
Splitting the rendering to each individual box makes it behave more like the current implementation, though modifying the containers' attributes now requires more code outside of the React implementation.

I made a quick prototype of this in https://github.com/dshlass/covid19japan/pull/1, but I didn't measure performance, and there seem to be some discrepancies in the layout (maybe changes in master...?). Just food for thought.