@dshlass 2020-08-07
> Splitting the rendering to each individual box makes it behave more like the current implementation, though modifying the containers' attributes now requires more code outside of the React implementation.
> 
> I made a quick prototype of this in [dshlass#1](https://github.com/dshlass/covid19japan/pull/1), but I didn't measure performance, and there seem to be some discrepancies in the layout (maybe changes in master...?). Just food for thought.

Wow thank you for this! It feels much like the live site does. I wanted to try something similar but couldn't wrap my head about it - great solution!

I noticed the discrepancies in master as well and I think they are due to the addition of Russian and Thai to the languages.

Attached are the previous lighthouse scores with the addition of your PR (afterSections.html)
[lighthouse.zip](https://github.com/reustle/covid19japan/files/5039187/lighthouse.zip)

Here is the attached performance differences
Summary:
**Speed Index:** 1.0s to 0.5s to **0.8s**
**Time to Interactive:** 8.4s to 6.5s to **4.1s**
**Total Blocking Time:** 5,810ms to 4,970ms to **2,910ms**
**Cumulative Layout Shift:** 0.03s to 0.014s to **0.095**

These lighthouse scores are pretty inconsistent but I agree that your PR make it feel the same as the live version.

I think that as the project slowly starts to migrate towards react there will be a lot more code being added outside of the React implementation. Once it gets closer to completion most of it can be removed and it can be a functioning React app with a single ReactDOM.render call.

This is my first time contributing to an open source project. Can I merge your PR into my branch without affecting anything else? I'd love to include your code in my PR.