@vlumi 2020-08-07
> I think that as the project slowly starts to migrate towards react there will be a lot more code being added outside of the React implementation. Once it gets closer to completion most of it can be removed and it can be a functioning React app with a single ReactDOM.render call.

I agree, the biggest turning point being when the data loading can be shifted to React, to allow the components to be drawn with default state before the data is ready.

> This is my first time contributing to an open source project. Can I merge your PR into my branch without affecting anything else? I'd love to include your code in my PR.

Of course, I created the pull request against your repo just so that there wouldn't be conflicting pull request against this main repo.