@stalkerg 2021-03-22
Probably Svelte will be much faster and simpler in this case, same as on the NYT tracker.